Agrego Tests y uso BigDecimal

+++ src/main/java/dds/monedero/model/Movimiento.java
@@ -1,21 +1,22 @@
 package dds.monedero.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Movimiento {
   private LocalDate fecha;
   //En ningún lenguaje de programación usen jamás doubles para modelar dinero en el mundo real
   //siempre usen numeros de precision arbitraria, como BigDecimal en Java y similares
-  private double monto;
+  private BigDecimal monto;
   private TipoOperacion tipoOperacion;
 
-  public Movimiento(LocalDate fecha, double monto, TipoOperacion tipoOperacion) {
+  public Movimiento(LocalDate fecha, BigDecimal monto, TipoOperacion tipoOperacion) {
     this.fecha = fecha;
     this.monto = monto;
     this.tipoOperacion = tipoOperacion;
   }
 
-  public double getMonto() {
+  public BigDecimal getMonto() {
     return monto;
   }
 
@@ -39,8 +40,7 @@ public class Movimiento {
     return tipoOperacion.isDeposito();
   }
 
-  public double calcularValor(double saldo, double monto) {
+  public BigDecimal calcularValor(BigDecimal saldo, BigDecimal monto) {
     return this.tipoOperacion.calcularNuevoSaldo(saldo,monto);
   }
-
 }
